migrate test to different class
