+++ scalding-parquet-scrooge/src/main/java/com/twitter/scalding/parquet/scrooge/ScroogeReadSupport.java
@@ -131,8 +131,8 @@ public class ScroogeReadSupport<T extends ThriftStruct> extends ThriftReadSuppor
   public static MessageType getSchemaForRead(MessageType fileMessageType, MessageType projectedMessageType) {
     assertGroupsAreCompatible(fileMessageType, projectedMessageType);
     return ParquetCollectionFormatForwardCompatibility.forwardCompatibleMessage(
-        projectedMessageType,
-        fileMessageType);
+        projectedMessageType, fileMessageType
+    );
   }
 
   /**
