+++ scalding-parquet-scrooge/src/test/scala/com/twitter/scalding/parquet/scrooge/ScroogeReadSupportTests.scala
@@ -7,7 +7,7 @@ import com.twitter.scalding.platform.{HadoopPlatformJobTest, HadoopSharedPlatfor
 import com.twitter.scalding.typed.TypedPipe
 import com.twitter.scalding.{Args, Job}
 import org.apache.parquet.io.InvalidRecordException
-import org.apache.parquet.schema.MessageTypeParser
+import org.apache.parquet.schema.{MessageTypeParser}
 import org.scalatest.{Matchers, WordSpec}
 
 class ScroogeReadSupportTests extends WordSpec with Matchers with HadoopSharedPlatformTest {
@@ -87,481 +87,6 @@ class ScroogeReadSupportTests extends WordSpec with Matchers with HadoopSharedPl
       an[InvalidRecordException] should be thrownBy {
         ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
       }
-    }
-  }
-
-  "ScroogeReadSupport resolving list format" should {
-    "resolve list legacy format: project x_tuple to legacy array" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group country_codes (LIST) {
-          |    repeated binary array (UTF8);
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated binary country_codes_tuple (UTF8);
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      // note optional of result, and field rename
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated binary array (UTF8);
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve list legacy format: project x_tuple to 3-level" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group country_codes (LIST) {
-          |    repeated group list {
-          |      required binary element (UTF8);
-          |    }
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated binary country_codes_tuple (UTF8);
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      // note optional of result, and field rename
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated group list {
-          |      required binary element (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve list legacy format: project nested x_tuple to nested legacy array" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group foo (LIST) {
-          |    repeated group array (LIST) {
-          |      repeated binary array (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group foo_tuple (LIST) {
-          |      repeated binary foo_tuple_tuple (UTF8);
-          |    }
-          |  }
-          |}
-      """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      // note optional of result, and field rename
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group array {
-          |      repeated binary array (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve list legacy format: project nested x_tuple to nested array" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group foo (LIST) {
-          |    repeated group array (LIST) {
-          |      repeated binary array (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group foo_tuple (LIST) {
-          |      repeated binary foo_tuple_tuple (UTF8);
-          |    }
-          |  }
-          |}
-      """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      // note optional of result, and field rename
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group array {
-          |      repeated binary array (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve list in group legacy format: project x_tuple to nested 3-level" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group foo (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated group list {
-          |          required binary element (UTF8);
-          |        }
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group foo_tuple (LIST) {
-          |      repeated binary foo_tuple_tuple (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated group list {
-          |          required binary element (UTF8);
-          |        }
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve: binary array to 3-level nesting" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group country_codes (LIST) {
-          |    repeated group list {
-          |      required binary element (UTF8);
-          |    }
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-
-      // inner list is `binary array`
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |     repeated binary array (UTF8);
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated group list {
-          |      required binary element (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-
-    }
-
-    "resolve: identity 3-level" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group country_codes (LIST) {
-          |    repeated group list {
-          |      required binary element (UTF8);
-          |    }
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, fileType)
-      projected shouldEqual fileType
-
-    }
-
-    "resolve nested list: project inner legacy array to 3-level nesting" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group array_of_country_codes (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated group list {
-          |          required binary element (UTF8);
-          |        }
-          |      }
-          |    }
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-
-      // inner list is `binary array`
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group array_of_country_codes (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated binary array (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group array_of_country_codes (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated group list {
-          |          required binary element (UTF8);
-          |        }
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve projected struct in list: repeated group element to 3-level nesting" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  required group country_codes (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        required binary foo (UTF8);
-          |        required binary bar (UTF8);
-          |        required binary zing (UTF8);
-          |      }
-          |    }
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated group element {
-          |      optional binary foo (UTF8);
-          |      required binary zing (UTF8);
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group country_codes (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        optional binary foo (UTF8);
-          |        required binary zing (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve standard 3-level list to 2-level" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message scalding_schema {
-          |  required group array_of_country_codes (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated binary array (UTF8);
-          |      }
-          |    }
-          |  }
-          |  required int32 x;
-          |}
-        """.stripMargin)
-
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group array_of_country_codes (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated group list {
-          |          required binary element (UTF8);
-          |        }
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group array_of_country_codes (LIST) {
-          |    repeated group list {
-          |      required group element (LIST) {
-          |        repeated binary array (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve list in group containing list" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  optional group connect_delays (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        optional binary description (UTF8);
-          |        optional binary created_by (UTF8);
-          |        optional group currencies (LIST) {
-          |          repeated group list {
-          |            required binary element (UTF8);
-          |          }
-          |        }
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group connect_delays (LIST) {
-          |    repeated group connect_delays_tuple {
-          |      optional binary description (UTF8);
-          |      optional group currencies (LIST) {
-          |        repeated binary currencies_tuple (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group connect_delays (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        optional binary description (UTF8);
-          |        optional group currencies (LIST) {
-          |          repeated group list {
-          |            required binary element (UTF8);
-          |          }
-          |        }
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
-    }
-
-    "resolve projection of different level" in {
-      val fileType = MessageTypeParser.parseMessageType(
-        """
-          |message spark_schema {
-          |  optional group foo (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        required binary zing (UTF8);
-          |        required binary bar (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val requestedProjection = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        optional binary bar (UTF8);
-          |        required binary zing (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      val projected = ScroogeReadSupport.getSchemaForRead(fileType, requestedProjection)
-      val expected = MessageTypeParser.parseMessageType(
-        """
-          |message SampleProjection {
-          |  optional group foo (LIST) {
-          |    repeated group list {
-          |      required group element {
-          |        optional binary bar (UTF8);
-          |        required binary zing (UTF8);
-          |      }
-          |    }
-          |  }
-          |}
-        """.stripMargin)
-      projected shouldEqual expected
     }
   }
